backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
elevenlabs==0.2.26
python-dotenv==1.0.0
ebooklib==0.18
beautifulsoup4==4.12.2
pydantic==2.4.2
pydantic-settings==2.0.3
SQLAlchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
fastapi-users[sqlalchemy]==12.1.2 